move json.py to avoid name collision
